Debouncer now u64-based. Pistons should work.

+++ src/debouncer.rs
@@ -1,5 +1,5 @@
 struct BitsWhere {
-    n: u16,
+    n: u64,
 }
 
 impl Iterator for BitsWhere {
@@ -21,14 +21,14 @@ impl Iterator for BitsWhere {
 }
 
 pub struct Debouncer<const DEPTH: usize> {
-    history: [u16; DEPTH],
+    history: [u64; DEPTH],
     index: usize,
-    stable_state: u16,
-    prev_stable_state: u16,
+    stable_state: u64,
+    prev_stable_state: u64,
 }
 
 impl<const DEPTH: usize> Debouncer<DEPTH> {
-    pub fn new(initial: u16) -> Self {
+    pub fn new(initial: u64) -> Self {
         Self {
             history: [initial; DEPTH],
             index: 0,
@@ -37,12 +37,12 @@ impl<const DEPTH: usize> Debouncer<DEPTH> {
         }
     }
 
-    pub fn update(&mut self, raw_reading: u16) {
+    pub fn update(&mut self, raw_reading: u64) {
         self.history[self.index] = raw_reading;
         self.index = (self.index + 1) % DEPTH;
 
-        let mut stable_high = !(0 as u16);
-        let mut stable_low = 0 as u16;
+        let mut stable_high = !(0 as u64);
+        let mut stable_low = 0 as u64;
 
         for val in self.history.iter() {
             stable_high &= *val;
